api/requirements.txt
fastapi
uvicorn[standard]
numpy
matplotlib
dwave-neal
pydantic
python-multipart
